+++ kernel/include/kernel_internal.h
@@ -34,8 +34,6 @@ void z_init_thread_base(struct _thread_base *thread_base, int priority,
 
 FUNC_NORETURN void z_cstart(void);
 
-void z_device_state_init(void);
-
 extern FUNC_NORETURN void z_thread_entry(k_thread_entry_t entry,
 			  void *p1, void *p2, void *p3);
