+++ kernel/init.c
@@ -38,7 +38,6 @@
 #include <kswap.h>
 #include <zephyr/timing/timing.h>
 #include <zephyr/logging/log.h>
-#include <zephyr/pm/device_runtime.h>
 #include <zephyr/internal/syscall_handler.h>
 #include <zephyr/arch/common/init.h>
 
@@ -193,49 +192,20 @@ extern volatile uintptr_t __stack_chk_guard;
 __pinned_bss
 bool z_sys_post_kernel;
 
-static int do_device_init(const struct device *dev)
+/* defined in device.c */
+extern int do_device_init(const struct device *dev);
+
+/**
+ * @brief Initialize state for all static devices.
+ *
+ * The state object is always zero-initialized, but this may not be
+ * sufficient.
+ */
+static void z_device_state_init(void)
 {
-	int rc = 0;
-
-	if (dev->ops.init != NULL) {
-		rc = dev->ops.init(dev);
-		/* If initialization failed, record in dev->state->init_res
-		 * the POSITIVE value of the resulting errno
-		 */
-		if (rc != 0) {
-			/* device's init function should return:
-			 *   0 on success
-			 *   a negative value on failure (-errno)
-			 * errno value maps to an uint8_t range as of now.
-			 */
-			__ASSERT(rc >= -UINT8_MAX && rc < 0, "device %s init: invalid error (%d)",
-				 dev->name, rc);
-
-			if (rc < 0) {
-				rc = -rc;
-			}
-			/* handle error value overflow in production
-			 * this is likely a bug in the device's init function. Signals it
-			 */
-			if (rc > UINT8_MAX) {
-				rc = UINT8_MAX;
-			}
-			dev->state->init_res = rc;
-		}
+	STRUCT_SECTION_FOREACH(device, dev) {
+		k_object_init(dev);
 	}
-
-	/* device initialization has been invoked */
-	dev->state->initialized = true;
-
-	if (rc == 0) {
-		/* Run automatic device runtime enablement */
-		(void)pm_device_runtime_auto_enable(dev);
-	}
-
-	/* here, the value of rc is either 0 or +errno
-	 * flip the sign to return a negative value on failure as expected
-	 */
-	return -rc;
 }
 
 /**
@@ -281,26 +251,7 @@ static void z_sys_init_run_level(enum init_level level)
 	}
 }
 
-
-int z_impl_device_init(const struct device *dev)
-{
-	if (dev->state->initialized) {
-		return -EALREADY;
-	}
-
-	return do_device_init(dev);
-}
-
-#ifdef CONFIG_USERSPACE
-static inline int z_vrfy_device_init(const struct device *dev)
-{
-	K_OOPS(K_SYSCALL_OBJ_INIT(dev, K_OBJ_ANY));
-
-	return z_impl_device_init(dev);
-}
-#include <zephyr/syscalls/device_init_mrsh.c>
-#endif
-
+/* defined in banner.c */
 extern void boot_banner(void);
